+++ Redoublet-backend/Redoublet-backend/Models/Player.cs
@@ -1,0 +1,11 @@
+﻿namespace Redoublet.Backend.Models
+{
+    public class Player
+    {
+        public string Name { get; set; }
+
+        public List<Card> Cards { get; set; }
+
+        public bool Vulnerable { get; set; }
+    }
+}
