+++ Redoublet-backend/Redoublet-backend/Models/StartGameRequest.cs
@@ -1,0 +1,10 @@
+﻿namespace Redoublet.Backend.Models
+{
+    public class StartGameRequest
+    {
+        public string NameNorth { get; set; }
+        public string NameEast { get; set; }
+        public string NameSouth { get; set; }
+        public string NameWest { get; set; }
+    }
+}
