+++ Redoublet-backend/Redoublet-backend/Models/Trick.cs
@@ -1,0 +1,9 @@
+﻿namespace Redoublet.Backend.Models
+{
+    public class Trick
+    {
+        public Card[] Cards { get; set; }
+
+        public Side Winner { get; set; }
+    }
+}
